+++ mario/__init__.py
@@ -43,6 +43,9 @@ Package dependencies:
 
 """
 
+
+import pandas as pd
+
 from mario.version import __version__
 from mario.core.AttrData import Database
 from mario.core.CoreIO import CoreModel
@@ -98,5 +101,6 @@ from mario.settings.settings import (
     Index,
 )
 from mario.tools.constants import IOT, SUT
+
 
 __authors__ = " 'Mohammad Amin Tahavori', Lorenzo Rinaldi', 'Nicolò Golinucci' "